Change setup/Mac.sh to be nicer on already setup systems
* Don't kill the Dock if that's already setup
* Fix checking for VS Code config file
* For now comment out the bits about setting up sudo for touch ID

+++ setup/Mac.sh
@@ -1,22 +1,24 @@
 #!/bin/bash
 
 # Show Cmd+tab app switcher on all displays, not just one with Dock
-defaults write com.apple.dock appswitcher-all-displays -bool true
-killall Dock
+if [ "$(defaults read com.apple.dock appswitcher-all-displays)" != 1 ]; then
+  defaults write com.apple.dock appswitcher-all-displays -bool true
+  killall Dock
+fi
 
 # Link VS Code settings into odd place it uses
 realname=$(perl -MCwd -e 'print Cwd::abs_path($ARGV[0])' "$0")
-dotfile_dir=$(dirname "$realname")
-
-defaults write com.microsoft.VSCode ApplePressAndHoldEnabled -bool false
+dotfile_dir=$(dirname "$(dirname "$realname")")
 
 vscode="$HOME/Library/Application Support/Code"
 mkdir -p "$vscode"
-[ -e "$vscode/settings.json" ] ||
+[ -e "$vscode/User/settings.json" ] ||
 	ln -is "$dotfile_dir/config/Code/User" "$vscode"
 
-cat <<EOF
-# Add following to top of /etc/pam.d/sudo
-auth       optional       /usr/local/lib/pam/pam_reattach.so
-auth       sufficient     pam_tid.so
-EOF
+defaults write com.microsoft.VSCode ApplePressAndHoldEnabled -bool false
+
+# cat <<EOF
+# # Add following to top of /etc/pam.d/sudo
+# auth       optional       /usr/local/lib/pam/pam_reattach.so
+# auth       sufficient     pam_tid.so
+# EOF
